+++ examples/plot_dataset_example.py
@@ -2,12 +2,12 @@
 #
 # License: BSD (3-clause)
 
-from braindecode.datasets.specific_datasets import BNCI2014001, TUHAbnormal
+from braindecode.datasets import MOABBDataset, TUHAbnormal
 
 # create a dataset based on BCIC IV 2a fetched with moabb
-ds = BNCI2014001(subject_ids=[4], trial_start_offset_samples=0,
-                 trial_stop_offset_samples=1000, supercrop_size_samples=1000,
-                 supercrop_stride_samples=1000)
+ds = MOABBDataset(dataset_name="BNCI2014001", subject_ids=[4],
+                  trial_start_offset_samples=0, trial_stop_offset_samples=1000,
+                  supercrop_size_samples=1000, supercrop_stride_samples=1000)
 
 # we can iterate through ds which yields an example x, target y,
 # and info as i_supercrop_in_trial, i_start_in_trial, and i_stop_in_trial
